fix: wasm32-wasi has been deprecated (#158)
The `wasm32-wasi` target has been renamed to `wasm32-wasip1`:

<https://blog.rust-lang.org/2024/04/09/updates-to-rusts-wasi-targets.html>

The `cargo-wasi` crate is also outdated and being replaced with
`cargo-component` for creating WASI components. It does not support the
new target name.
<https://github.com/bytecodealliance/cargo-wasi/issues/143>

However, this project doesn't currently define a WASI component, so we
can't use cargo-component to build or run tests. We do want to verify
that the project builds correctly for the WASM targets, so this change
runs the compiled test targets directly with `wasmtime`.